feature: changes to scale of depth to meters (conventional in other pipeline).

+++ gSLICr_Lib/engines/gSLICr_seg_engine_shared.h
@@ -92,7 +92,7 @@ _CPU_AND_GPU_CODE_ inline void init_cluster_centers_shared(const gSLICr::Vector4
 	out_spixel[cluster_idx].no_pixels = 0;
 }
 
-_CPU_AND_GPU_CODE_ inline void init_cluster_centers_shared(const gSLICr::Vector4f* inimg, const short* indepth, gSLICr::objects::spixel_d_info* out_spixel, gSLICr::Vector2i map_size, gSLICr::Vector2i img_size, int spixel_size, int x, int y)
+_CPU_AND_GPU_CODE_ inline void init_cluster_centers_shared(const gSLICr::Vector4f* inimg, const float* indepth, gSLICr::objects::spixel_d_info* out_spixel, gSLICr::Vector2i map_size, gSLICr::Vector2i img_size, int spixel_size, int x, int y)
 {
 	int cluster_idx = y * map_size.x + x;
 
@@ -126,7 +126,7 @@ _CPU_AND_GPU_CODE_ inline float compute_slic_distance(const gSLICr::Vector4f& pi
 	return sqrtf(retval);
 }
 
-_CPU_AND_GPU_CODE_ inline float compute_slic_distance(const gSLICr::Vector4f& pix, const short d, int x, int y, const gSLICr::objects::spixel_d_info& center_info, float wXY, float wD, float normalizer_xy, float normalizer_color, float normalizer_depth)
+_CPU_AND_GPU_CODE_ inline float compute_slic_distance(const gSLICr::Vector4f& pix, const float d, int x, int y, const gSLICr::objects::spixel_d_info& center_info, float wXY, float wD, float normalizer_xy, float normalizer_color, float normalizer_depth)
 {
 	float dcolor = (pix.x - center_info.color_info.x)*(pix.x - center_info.color_info.x)
 				 + (pix.y - center_info.color_info.y)*(pix.y - center_info.color_info.y)
@@ -175,7 +175,7 @@ _CPU_AND_GPU_CODE_ inline void find_center_association_shared(const gSLICr::Vect
 	if (minidx >= 0) out_idx_img[idx_img] = minidx;
 }
 
-_CPU_AND_GPU_CODE_ inline void find_center_association_shared(const gSLICr::Vector4f* inimg, const short* indep, const gSLICr::objects::spixel_d_info* in_spixel_map, int* out_idx_img, gSLICr::Vector2i map_size, gSLICr::Vector2i img_size, int spixel_size, float wXY, float wD, int x, int y, float max_xy_dist, float max_color_dist, float max_depth_dist)
+_CPU_AND_GPU_CODE_ inline void find_center_association_shared(const gSLICr::Vector4f* inimg, const float* indep, const gSLICr::objects::spixel_d_info* in_spixel_map, int* out_idx_img, gSLICr::Vector2i map_size, gSLICr::Vector2i img_size, int spixel_size, float wXY, float wD, int x, int y, float max_xy_dist, float max_color_dist, float max_depth_dist)
 {
 	int idx_img = y * img_size.x + x;
